+++ app/models/collecting_event/georeference.rb
@@ -109,7 +109,7 @@ module CollectingEvent::Georeference
     when :verbatim_map_center
       verbatim_map_center
     when :geographic_area
-      geographic_area.default_geographic_item.geo_object.centroid
+      geographic_area.default_geographic_item.centroid
     else
       nil
     end